# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChargeGrid Intelligence API",
...
    "contact": {
      "name": "ChargeGrid Intelligence",
      "url": "https://www.goodwe.com/"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/campaigns": {
      "get": {
        "tags": [
          "campanhas"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_campaigns_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "admin: escolhe o site da rede",
              "title": "Site Id"
            },
            "description": "admin: escolhe o site da rede"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampanhaOut"
                  },
                  "title": "Response Listar Api V1 Campaigns Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "campanhas"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_campaigns_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "admin: escolhe o site da rede",
              "title": "Site Id"
            },
            "description": "admin: escolhe o site da rede"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampanhaIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampanhaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campanha_id}": {
      "patch": {
        "tags": [
          "campanhas"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_campaigns__campanha_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campanha_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campanha Id"
            }
          },
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "admin: escolhe o site da rede",
              "title": "Site Id"
            },
            "description": "admin: escolhe o site da rede"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampanhaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "campanhas"
        ],
        "summary": "Encerrar",
        "description": "Encerra a campanha. NAO apaga.\n\nApagar levaria junto o progresso de quem estava no meio dela, e o RESTRICT em\n`rewards.campaign_id` recusaria a exclusao assim que a primeira recompensa\ntivesse sido concedida - com um erro de banco que ninguem sabe ler. Desativar\ne' o que o operador quer dizer com \"encerrar\", e preserva o rastro.",
        "operationId": "encerrar_api_v1_campaigns__campanha_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campanha_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campanha Id"
            }
          },
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "admin: escolhe o site da rede",
              "title": "Site Id"
            },
            "description": "admin: escolhe o site da rede"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campanha_id}/desempenho": {
      "get": {
        "tags": [
          "campanhas"
        ],
        "summary": "Desempenho",
        "operationId": "desempenho_api_v1_campaigns__campanha_id__desempenho_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campanha_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campanha Id"
            }
          },
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "admin: escolhe o site da rede",
              "title": "Site Id"
            },
            "description": "admin: escolhe o site da rede"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesempenhoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "CampanhaIn": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 120,
            "title": "Nome"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "patrocinador": {
            "type": "string",
            "title": "Patrocinador",
            "default": "rede"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ends At"
          },
          "ativa": {
            "type": "boolean",
            "title": "Ativa",
            "default": true
          },
          "beneficio_tipo": {
            "type": "string",
            "title": "Beneficio Tipo"
          },
          "beneficio_valor": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Beneficio Valor"
          },
          "teto_por_recompensa": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Teto Por Recompensa"
          },
          "orcamento_brl": {
            "type": "number",
            "minimum": 0.0,
            "title": "Orcamento Brl",
            "default": 0
          },
          "limite_por_motorista": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Limite Por Motorista"
          },
          "missoes": {
            "items": {
              "$ref": "#/components/schemas/MissaoIn"
            },
            "type": "array",
            "title": "Missoes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "nome",
          "starts_at",
          "ends_at",
          "beneficio_tipo",
          "beneficio_valor"
        ],
        "title": "CampanhaIn"
      },
      "CampanhaOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "patrocinador": {
            "type": "string",
            "title": "Patrocinador"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "fleet_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fleet Id"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ends At"
          },
          "ativa": {
            "type": "boolean",
            "title": "Ativa"
          },
          "beneficio_tipo": {
            "type": "string",
            "title": "Beneficio Tipo"
          },
          "beneficio_valor": {
            "type": "number",
            "title": "Beneficio Valor"
          },
          "teto_por_recompensa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teto Por Recompensa"
          },
          "orcamento_brl": {
            "type": "number",
            "title": "Orcamento Brl"
          },
          "consumido_brl": {
            "type": "number",
            "title": "Consumido Brl"
          },
          "limite_por_motorista": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limite Por Motorista"
          },
          "missoes": {
            "items": {
              "$ref": "#/components/schemas/MissaoOut"
            },
            "type": "array",
            "title": "Missoes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "descricao",
          "patrocinador",
          "site_id",
          "fleet_id",
          "starts_at",
          "ends_at",
          "ativa",
          "beneficio_tipo",
          "beneficio_valor",
          "teto_por_recompensa",
          "orcamento_brl",
          "consumido_brl",
          "limite_por_motorista"
        ],
        "title": "CampanhaOut",
        "description": "Visao do operador: inclui o dinheiro."
      },
      "DesempenhoOut": {
        "properties": {
          "campanha_id": {
            "type": "string",
            "format": "uuid",
            "title": "Campanha Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "motoristas_alcancados": {
            "type": "integer",
            "title": "Motoristas Alcancados"
          },
          "missoes_concluidas": {
            "type": "integer",
            "title": "Missoes Concluidas"
          },
          "recompensas_creditadas": {
            "type": "integer",
            "title": "Recompensas Creditadas"
          },
          "consumido_brl": {
            "type": "number",
            "title": "Consumido Brl"
          },
          "orcamento_brl": {
            "type": "number",
            "title": "Orcamento Brl"
          },
          "orcamento_disponivel": {
            "type": "number",
            "title": "Orcamento Disponivel"
          },
          "percentual_consumido": {
            "type": "number",
            "title": "Percentual Consumido"
          }
        },
        "type": "object",
        "required": [
          "campanha_id",
          "nome",
          "motoristas_alcancados",
          "missoes_concluidas",
          "recompensas_creditadas",
          "consumido_brl",
          "orcamento_brl",
          "orcamento_disponivel",
          "percentual_consumido"
        ],
        "title": "DesempenhoOut",
        "description": "O que o operador precisa para decidir se a campanha vale.\n\nSem `motoristas_alcancados` e `concluidas`, \"consumido_brl\" sozinho nao diz\nse o dinheiro comprou comportamento ou so' saiu do caixa."
      },
...
      "MissaoIn": {
        "properties": {
          "codigo": {
            "type": "string",
            "maxLength": 40,
            "title": "Codigo"
          },
          "titulo": {
            "type": "string",
            "maxLength": 120,
            "title": "Titulo"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "metrica": {
            "type": "string",
            "title": "Metrica"
          },
          "alvo": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Alvo"
          },
          "janela": {
            "type": "string",
            "title": "Janela",
            "default": "campanha"
          },
          "ordem": {
            "type": "integer",
            "title": "Ordem",
            "default": 0
          },
          "repetivel": {
            "type": "boolean",
            "title": "Repetivel",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "codigo",
          "titulo",
          "metrica",
          "alvo"
        ],
        "title": "MissaoIn"
      },
      "MissaoOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "metrica": {
            "type": "string",
            "title": "Metrica"
          },
          "alvo": {
            "type": "number",
            "title": "Alvo"
          },
          "janela": {
            "type": "string",
            "title": "Janela"
          },
          "ordem": {
            "type": "integer",
            "title": "Ordem"
          },
          "repetivel": {
            "type": "boolean",
            "title": "Repetivel"
          }
        },
        "type": "object",
        "required": [
          "id",
          "codigo",
          "titulo",
          "descricao",
          "metrica",
          "alvo",
          "janela",
          "ordem",
          "repetivel"
        ],
        "title": "MissaoOut"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  },
...
}